src/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
PyGithub
python-multipart
sentence-transformers
torch
numpy